Update wdi config
