+++ .github/workflows/tests.yaml
@@ -5,6 +5,7 @@ on:
 
 jobs:
   tests:
+    name: Tests
     runs-on: ubuntu-latest
     steps:
       - name: Start Selenoid server
