+++ package.json
@@ -13,8 +13,8 @@
     "playground": "npm start --workspace=@cere-games-sdk/playground",
     "build": "npm run build --workspaces --if-present",
     "lint": "eslint --quiet '**/*.{ts,tsx,js,jsx}'",
-    "test": "npm run test --workspaces --if-present",
-    "test:ci": "npm run test:ci --workspaces --if-present",
+    "test": "npm test --workspaces --if-present -- $*",
+    "test:ci": "npm run test:ci --workspaces --if-present -- $*",
     "sdk:publish": "npm publish --workspace=@cere/games-sdk-ui --workspace=@cere/games-sdk --access public"
   },
   "workspaces": [
