--- tests/wdio/browsers.json
@@ -1,12 +1,0 @@
-{
-  "chrome": {
-    "default": "latest",
-    "versions": {
-      "latest": {
-        "image": "selenoid/chrome:112.0",
-        "port": "4444",
-        "path": "/"
-      }
-    }
-  }
-}
