+++ tests/wdio/config/ci.ts
@@ -1,19 +1,16 @@
 import { config as baseConfig } from './base';
-import { chromeCapability } from './chromeDriver';
-import { options } from './options';
 
 export const config: WebdriverIO.Config = {
   ...baseConfig,
 
+  /**
+   * The hardcoded IP address is the address of GitHub actions host
+   *
+   * TODO: Figure out a better way to dected CI host IP, insted of hardcoding this one
+   */
   baseUrl: 'http://172.17.0.1:4567/examples/',
+
   hostname: 'localhost',
   port: 4444,
   path: '/wd/hub/',
-
-  capabilities: [
-    {
-      ...chromeCapability,
-      maxInstances: options.maxInstances,
-    },
-  ],
 };
