+++ tests/wdio/config/index.ts
@@ -1,8 +1,3 @@
-import { options } from './options';
-
-const variants: Record<typeof options.variant, WebdriverIO.Config> = {
-  chrome: require('./chromeDriver').config,
-  selenoid: require('./selenoid').config,
-};
-
-export const config = variants[options.variant || 'chrome'];
+export * from './options';
+export { config as ci } from './ci';
+export { config as local } from './local';
