+++ tests/wdio/config/options.ts
@@ -2,7 +2,7 @@ import { hideBin } from 'yargs/helpers';
 import yargs from 'yargs';
 
 export const options = yargs(hideBin(process.argv)).options({
-  variant: { choices: ['selenoid', 'chrome'] as const },
+  ci: { type: 'boolean', default: false },
   headless: { type: 'boolean', default: false },
   maxInstances: { type: 'number', default: 1 },
 }).argv;
