+++ tests/wdio/wdio.conf.ts
@@ -1,1 +1,3 @@
-export * from './config';
+import { options, ci, local } from './config';
+
+export const config = options.ci ? ci : local;
